Lambda/MochiBot_Entry/requirements.txt
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
idna==3.10
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.4
urllib3==2.5.0
